@@ -3,37 +3,40 @@
   "version": "0.8.1",
   "private": true,
   "devDependencies": {
+    "@types/bun": "1.0.12",
     "@types/fluent-ffmpeg": "2.1.22",
-    "@types/jest": "^29.5.5",
+    "@types/jest": "29.5.5",
     "@types/node": "20.8.0",
     "@types/prettier": "2.7.3",
     "@types/react": "18.2.23",
     "@types/resize-observer-browser": "0.1.8",
     "@types/webmidi": "2.0.8",
     "autoprefixer": "10.4.16",
-    "bun-types": "1.0.3",
+    "eslint-config-next": "14.1.4",
     "eslint": "8.56.0",
-    "eslint-config-next": "14.0.4",
     "fluent-ffmpeg": "2.1.2",
     "jsdom": "22.1.0",
     "postcss": "8.4.31",
     "prettier": "3.0.3",
+    "sharp": "0.33.3",
     "skia-canvas": "1.0.1",
-    "tailwindcss": "3.3.3",
+    "tailwindcss": "3.4.3",
     "ts-node": "10.9.1",
     "tsconfig-paths": "4.2.0",
-    "typescript": "5.2.2"
+    "typescript": "5.4.4"
   },
   "dependencies": {
-    "@preact/signals-react": "1.3.6",
     "@radix-ui/react-slider": "1.1.2",
     "@radix-ui/react-tooltip": "1.0.7",
     "@tonejs/midi": "2.0.28",
     "clsx": "2.0.0",
-    "next": "13.5.3",
-    "react": "18.2.0",
+    "jotai-effect": "0.3.2",
+    "jotai": "2.6.2",
+    "next": "14.1.4",
     "react-dom": "18.2.0",
-    "react-feather": "2.0.10"
+    "react-feather": "2.0.10",
+    "react": "18.2.0",
+    "swr": "2.2.5"
   },
   "scripts": {
     "start": "next start",
