@@ -1,5 +1,5 @@
 import { useSong } from '@/features/data'
-import Player from '@/features/player'
+import { usePlayer } from '@/features/player'
 import { getHandSettings, SongVisualizer } from '@/features/SongVisualization'
 import { SongSource } from '@/types'
 import { useEffect } from 'react'
@@ -9,12 +9,11 @@ interface SongPreviewProps {
   songBytes?: ArrayBuffer
   songId: string
   source: SongSource
-  className?: string
 }
 
-function SongPreview({ songId, source, className }: SongPreviewProps) {
-  const player = Player.player()
+function SongPreview({ songId, source }: SongPreviewProps) {
   const { data: song, error } = useSong(songId, source)
+  const player = usePlayer()
 
   useEffect(() => {
     if (!song) {
@@ -28,7 +27,7 @@ function SongPreview({ songId, source, className }: SongPreviewProps) {
     <SongVisualizer
       song={song}
       config={songConfig}
-      getTime={() => Player.player().getTime()}
+      getTime={() => player.getTime()}
       hand="both"
       handSettings={getHandSettings(songConfig)}
     />
