@@ -5,6 +5,7 @@ import { useRef, useMemo } from 'react'
 import * as touchscroll from '@/features/SongVisualization/touchscroll'
 import { PIXELS_PER_SECOND as pps } from './utils'
 import { Canvas } from '@/components'
+import { usePlayer } from '../player'
 
 type HandSettings = {
   [trackId: string]: {
@@ -37,6 +38,7 @@ function CanvasRenderer({
 }: CanvasRendererProps) {
   const { width, height, measureRef } = useSize()
   const canvasRef = useRef<HTMLCanvasElement>()
+  const player = usePlayer()
 
   const canvasRect: DOMRect = useMemo(() => {
     return canvasRef.current?.getBoundingClientRect() ?? {}
@@ -66,6 +68,7 @@ function CanvasRenderer({
       canvasRect,
       selectedRange,
       game,
+      player,
     }
     render(state)
   }
@@ -74,9 +77,9 @@ function CanvasRenderer({
     <div
       className="absolute w-full h-full touch-none"
       ref={measureRef}
-      onPointerMove={(e) => enableTouchscroll && touchscroll.handleMove(e.nativeEvent)}
-      onPointerDown={(e) => enableTouchscroll && touchscroll.handleDown(e.nativeEvent)}
-      onPointerUp={(e) => enableTouchscroll && touchscroll.handleUp(e.nativeEvent)}
+      onPointerMove={(e) => enableTouchscroll && touchscroll.handleMove(player, e.nativeEvent)}
+      onPointerDown={(e) => enableTouchscroll && touchscroll.handleDown(player, e.nativeEvent)}
+      onPointerUp={(e) => enableTouchscroll && touchscroll.handleUp(player, e.nativeEvent)}
     >
       <Canvas ref={canvasRef} render={renderCanvas} />
     </div>
