@@ -1,17 +1,18 @@
-import Player from '@/features/player'
 import { clamp } from '@/utils'
 import { getPointerVelocity } from '../pointer'
 import { intersectsWithPiano } from './falling-notes'
 import { PIXELS_PER_SECOND as pps } from './utils'
+import { getDefaultStore } from 'jotai'
+import { Player } from '../player'
 
-const player = Player.player()
+type JotaiStore = ReturnType<typeof getDefaultStore>
 
 let isDragging_ = false
 export function isDragging(): boolean {
   return isDragging_
 }
 
-function seekSeconds(seconds: number) {
+function seekSeconds(player: Player, seconds: number) {
   const songTime = clamp(seconds + player.getTime(), {
     min: 0,
     max: player.getDuration(),
@@ -21,24 +22,24 @@ function seekSeconds(seconds: number) {
 
 const decayRate = 0.96
 let acceleration = 0
-function decay() {
+function decay(player: Player) {
   isDragging_ = false
   requestAnimationFrame(() => {
     const songSeconds = acceleration / pps
     // End touchscroll when the acceleration catches up to the natural song velocity.
-    const endSnap = (1 / 60) * player.bpmModifier.value // TODO: instead of 1/60 it should be 1 / frameRate.
+    const endSnap = (1 / 60) * player.store.get(player.getBpmModifier()) // TODO: instead of 1/60 it should be 1 / frameRate.
     if (Math.abs(songSeconds) > endSnap) {
-      seekSeconds(songSeconds)
+      seekSeconds(player, songSeconds)
       acceleration *= decayRate
-      decay()
+      decay(player)
     } else {
-      endInertialScroll()
+      endInertialScroll(player)
     }
   })
 }
 
 let wasPlaying = false
-export function handleDown(e: PointerEvent) {
+export function handleDown(player: Player, e: PointerEvent) {
   if (intersectsWithPiano(e.clientY)) {
     isDragging_ = false
     return
@@ -55,14 +56,14 @@ export function handleDown(e: PointerEvent) {
   // TODO: doubleclick pause / play
 }
 
-export function handleUp(e: PointerEvent) {
+export function handleUp(player: Player, e: PointerEvent) {
   const target = e.target as HTMLDivElement
   target.releasePointerCapture(e.pointerId)
-  decay()
+  decay(player)
 }
 
 // Should resume playback.
-function endInertialScroll() {
+function endInertialScroll(player: Player) {
   if (wasPlaying) {
     wasPlaying = false
     player.play()
@@ -70,12 +71,12 @@ function endInertialScroll() {
   acceleration = 0
 }
 
-export function handleMove(e: PointerEvent) {
+export function handleMove(player: Player, e: PointerEvent) {
   if (!isDragging_) {
     return
   }
   const yVel = getPointerVelocity().y
-  seekSeconds(yVel / pps)
+  seekSeconds(player, yVel / pps)
 
   //  Threshold to prevent accidental flings
   const threshold = 5
