@@ -1,8 +1,9 @@
+'use client'
 import React, { useState } from 'react'
 import clsx from 'clsx'
 import { Select } from '@/components'
 import { LeftHand, RightHand, Volume2, VolumeX } from '@/icons'
-import Player from '@/features/player'
+import { usePlayer } from '@/features/player'
 import { InstrumentName, gmInstruments } from '@/features/synth'
 import { Song, SongConfig, TrackSetting } from '@/types'
 import { formatInstrumentName } from '@/utils'
@@ -47,7 +48,7 @@ type SynthState = { error: boolean; loading: boolean }
 
 function InstrumentCard({ track, trackId, setTrack, noteCount }: CardProps) {
   const [synthState, setSynthState] = useState<SynthState>({ error: false, loading: false })
-  const player = Player.player()
+  const player = usePlayer()
 
   const handleSelectInstrument = (instrument: InstrumentName) => {
     setSynthState({ error: false, loading: true })
