@@ -1,40 +1,18 @@
-import Player from '@/features/player'
+import { usePlayer } from '@/features/player'
 import { useMemo } from 'react'
+import { useAtomValue } from 'jotai'
 
-type PlayerDispatcher = {
-  play: () => void
-  pause: () => void
-  toggle: () => void
-  stop: () => void
-}
-
-type PlayerStateHookReturn = [
-  { canPlay: boolean; playing: boolean; paused: boolean },
-  PlayerDispatcher,
-]
+type PlayerStateHookReturn = { canPlay: boolean; playing: boolean; paused: boolean }
 
 export default function usePlayerState(): PlayerStateHookReturn {
-  const player = Player.player()
-  const state = player.state.value
-
-  const dispatcher: PlayerDispatcher = useMemo(
-    () => ({
-      play: () => player.play(),
-      pause: () => player.pause(),
-      toggle: () => player.toggle(),
-      stop: () => player.stop(),
-    }),
-    [player],
-  )
+  const player = usePlayer()
+  const state = useAtomValue(player.state)
 
   return useMemo(() => {
-    return [
-      {
-        canPlay: state !== 'CannotPlay',
-        playing: state === 'Playing',
-        paused: state === 'Paused',
-      },
-      dispatcher,
-    ]
-  }, [state, dispatcher])
+    return {
+      canPlay: state !== 'CannotPlay',
+      playing: state === 'Playing',
+      paused: state === 'Paused',
+    }
+  }, [state])
 }
